Adapting the Settings class
Create a settings class that is acting like a singleton and prodives
access to values in the app.config file. This file should be able to
handle settings in the Rhenus.Service config group only, leaving the
settings in other settings groups to the implementing classes of
modules. This way, the config file can contain the settings for all
modules in separate groups.