Load assemblies dynamically
When the service is started and the user has specified a number of
modules in sections of the config file, the service should be finding
out about the module name from the section name. These names should be
printed to the Console for the moment but are required to load the
specified assemblies lateron.