Change target platform
A lot of things are not working with the .NET framework 2.0. This
includes automatically created properties and support for certain types.
Set the target framework to Mono / .NET 3.5 to avoid these problems. Any
fresh installation of the service should be running on a system that is
using this target platform anyways. If necessary the downwards
compatibility can be acchieved at a later time, but with much more
effort.