Add a deployment step to the build
Building the artifacts with the buildscript is nice, but we are going to
automate the publication and distribution, too. This requires someone to
take care about creating the folder structure to deploy and to copy the
latest binaries there. In a further step there should be a startup file
for sh on Linux and a bat on Windows that is able to step up the service
to run on that machine. When everything is in place, the deployment
folder should be zipped and published.