Implement the ServiceBase
The service is going to run as a Windows Service or a Daemon on Unix
Systems. Both are basing their services on the
System.ServiceProcess.ServiceBase class that has to be implemented to
run the process as a service. The specific methods that are defining
initialization and shutdown have to be implemented.