Watching out for the state of the service
Create a struct or class that is able to inform anyone that is
interested about the current services internal state. This make health
monitoring a lot easier, because the service knows if it has caught an
unhandled exception and can tell others about it. It should also be able
to tell about the current workload in the future.