Implement periodic tasks
In order to have tasks running more that once at a specified interval,
implement a scheduling method that return a handle to the task and make
it execute at the specified interval. This is part of the ITaskScheduler
and the periodic task should also be hidden behind an interface for
later easy changing.